requirements.txt
Google-Images-Search==1.3.9
discord.py==1.7.1